Add scan command
This commit adds bloody simple `scan` command as "base code" for the other commands.
The resulting JSON is not complete: paths are absolute (should be relative from root
path), copyright information is totally empty. However, as the output schema is still
unstable, I think this is okay quality to commit. The scan logic and file-system
related things are likely to change in the future works.